refactor(snowman): extract physics kernel (R3.10)
